renamed merger and downloader
